+++ src/main/java/me/grayingout/bot/listeners/interactions/WarningsListInteractionListeners.java
@@ -1,7 +1,7 @@
-package me.grayingout.bot.events.interactions;
+package me.grayingout.bot.listeners.interactions;
 
-import me.grayingout.database.accessor.DatabaseAccessorManager;
-import me.grayingout.database.objects.MemberWarningsListMessage;
+import me.grayingout.database.accessors.DatabaseAccessorManager;
+import me.grayingout.database.entities.MemberWarningsListMessage;
 import me.grayingout.util.Warnings;
 import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
@@ -9,7 +9,7 @@ import net.dv8tion.jda.api.hooks.ListenerAdapter;
 /**
  * Handles the user interacting with a warnings list
  */
-public final class WarningsListInteractionHandler extends ListenerAdapter {
+public final class WarningsListInteractionListeners extends ListenerAdapter {
 
     @Override
     public void onButtonInteraction(ButtonInteractionEvent event) {
