+++ src/main/java/me/grayingout/database/accessors/LevellingDatabaseAccessor.java
@@ -1,4 +1,4 @@
-package me.grayingout.database.accessor;
+package me.grayingout.database.accessors;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -9,8 +9,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
-import me.grayingout.database.objects.GuildLevelRole;
-import me.grayingout.database.objects.GuildMemberLevelExperience;
+import me.grayingout.database.entities.GuildLevelRole;
+import me.grayingout.database.entities.GuildMemberLevelExperience;
 import me.grayingout.database.query.DatabaseQuery;
 import me.grayingout.util.Levelling;
 import net.dv8tion.jda.api.entities.Guild;
