+++ src/main/java/me/grayingout/database/entities/GuildLevelRole.java
@@ -1,4 +1,4 @@
-package me.grayingout.database.objects;
+package me.grayingout.database.entities;
 
 import java.util.Comparator;
